Add ext_qt patch for Windows to support fractional DPI scaling
This adds a series of patches related to QTBUG-53022 [1] for Windows to
allow enabling support for fractional DPI scaling.

These series of patches adds the option
`QGuiApplication::setHighDpiScaleFactorRoundingPolicy` to control how
the DPI scale factor is rounded (or not), which can also be overridden
by the environment variable `QT_SCALE_FACTOR_ROUNDING_POLICY`. The
options include:

- Round: Round up for .5 and above.
- Ceil: Always round up.
- Floor: Always round down.
- RoundPreferFloor: Round up for .75 and above.
- PassThrough: Don't round.

Currently the default option is explicitly set to `Round` to obtain the
old behaviour (where 1.5x is rounded to 2x).

Ref:
- https://bugreports.qt.io/browse/QTBUG-53022
- https://codereview.qt-project.org/157173
- https://codereview.qt-project.org/157174
- https://codereview.qt-project.org/161334
- https://codereview.qt-project.org/163219
- https://codereview.qt-project.org/176381
- https://codereview.qt-project.org/209886