Add a wrapper for an updater for subtasks
This commit adds two classes that together make it easier to handle
calculation of progress in case of subtasks.
For example if you have several functions that do long, but easy
to estimate tasks like doing something for every layer or other item
from a collection, you can just put KisUpdaterWrapperSubtask
as an argument and then setProgress that is only relevant in that
function (from 0 to 100), not worrying about other subtasks.

This is in replacement of KoProgressUpdater (despite a bit different
usage) because the class in question doesn't work correctly
in threads without an event loop.