Bugfix: layers with alpha channel disabled correctly export as "svg:src-atop" for ORA
A slight logical bug in a series of if/else statements lead to the default "src-over". Attribute being applied to layers with alpha channel disabled, which should have been exported with "svg:src-atop". I fixed this by moving checks for non-composite-op types to the end, just before the check for "src-over".

An alternative solution would be to default to exporting "src-over" in the else statement instead of checking for it explicitly, and only checking for official ORA spec attributes.

(cherry picked from commit 8fb97635e9dd41dd05205d92a915077f35f8ab0f)