WIP: Support vector layers in ORA files
This is an implementation proposal for including support for vector layers
in the export and import of ORA files, pending further discussion with
the users and maintainers of other ORA compatible software programs.

In this implementation, a separate stack xml tag 'vector' is used in place
of 'layer' to indicate that the source should be interpreted as an SVG  file
instead of a PNG file. Programs which search the tree for <layer> tags should
naturally ignore loading these layers, which is consistent with the current
behavior of ORA import / export.

Alternatively, <layer> stack elements may be duel purpose, perhaps including
both a "src" and "vector-src" attribute, and both a PNG and SVG can be exported,
which would allow the vector layers to be loaded as raster layers in programs
that have no vector implementation. However, this would require a larger storage
size for these files.

There are many other possibilities too, I am ready for an extended discussion.

Additionally, this is one of my earlier contributions to Krita (and C++
projects in general), and I am interested to hear any feedback about my
design choices in the implementation.