Fix deform brush by making KoColorMixOp handle non-255 weights.
This fixes the bug where this brush slowly outputs everything to transparent.

The bug was caused by the fact that to sample the input pixels, the crossdevice
color picker was used, which in turn got access to subpixel data by utilizing
the randomsubaccessor, which was getting that data by mixing neighbouring pixels
via the KoColorMixOp. However, the randomsubaccessor isn't able to generate
perfect sums-to-255 weights every single time, meaning transparency was added.

This commit fixes that by making the weights in KoColorMixOp always use the sum
of the weights as the normalization factor.

Special thanks to Dmitry for figuring this out.

CCBUG:290383