Fix resetting seexpr script when renaming
Before this commit, if the user had some changes in the script and then
decide to rename the script, it would reset to the original script
saved into the preset.
This commit ensures that the script with changes is restored after
renaming process so the user doesn't lose any changes.
(This is the same behaviour as in brush presets editor).