Add sanitizing of document title
Before this commit, you could use any string for a document title.
This commit ensures that user won't be able to use / or \
in their document title.

BUG:429925

Examples of behaviour before:
* A. document title: `test`
Correct file.
Layers etc. were located in the archive in directory called `test`
* B. document title: `test/`
There was no way to open the file later correctly without
editing the metadata file (maindoc.xml or documentinfo.xml,
not sure because I haven't checked) because Krita would
look for layers in `test//layers/` instead of `test/layers/`.
(This issue always happened for document titles ending with `\`)
* C. document title: `testA/testB`
Correct file.
Layers etc. were located in the archive in a directory called
`testB` which was located in a directory called `testA`.
* D. document title: `../../../test`
Correct file.
Layers etc. were located in the path `../../../test`, which means
in `test` located in directory called `..` in `..` in `..`.

After this fix, the files that were still correctly opened before this
change (files with document titles with `/` but not having `/` at the
end) are still correctly opened, and saved if the user doesn't change
the document title.

This fix only prevents creating new files with those document titles
because of both the possibility of creating an incorrect file and
the unnecessary nesting of directories.