Bugfix: Keep canvas position on show canvas only
This patch checks the position of the mdiArea before and after
activating 'show canvas only', 'fullscreen', or 'toggle dockers' mode
and moves any subwindow that is not maximized to the same absolute
global position. Any subwindow that is maximized, or in tab mode,
the scrollbar is adjusted to the same position.

It also sets up an overlay to help reduce flickering of the screen
during the transition. The timeout can be adjusted or disabled by
the user.

Note: Any maximized subwindow or tab that is set to a non-constant
zoom mode(ex. Fit Page) will be ignored. As it is assumed the user
wishes to resize the canvas to fit.

Fixes Bug#392225