Impl. KisQuickWidgetCanvas to render QtQuick2 on canvas
This is a modification of KisOpenGLCanvas2 which uses
QQuickRenderControl to render QtQuick2 scenes on canvas. The difference
with this implementation and QQuickWidget is that, QQuickWidget uses a
timer with 5ms timeout to delay sync and render, which causes the
rendering of the QtQuick scene to be out of sync with QWidget updates,
making the canvas noticeably slower, or even halving its update rate
when vsync is in effect. This implementation instead syncs and renders
directly onto the FBO of the QOpenGLWidget in sync with paintEvent /
paintGL, successfully eliminating the issues with QQuickWidget.

Note that this does not implement forwarding input events to the
offscreen QQuickWindow. This will be left for a later time.