Teach canvas to track canvas updates separately
This allows us to redraw image tiles to the canvas FBO only when the
canvas projection has changed. This applies to the OpenGL canvas when
the canvas FBO is being used.

Note: When KisCanvas2::updateCanvas is called, it still assumes the
canvas projection needs updating because it cannot be sure. To benefit
from the new behaviour, one shall use the other updateCanvas** methods
instead.