Differenciate between enabled/visible state of the outline in KisToolPaint
Before there was only a "setOutlineEnabled" method that allowed to show/hide the
tool outline. This was used for example to hide the outline if the alternate
action (color sampling) was activated and show it when the action was deactivated.
The problem with this is that if a KisToolPaint derived class wants to always hide
the outline via "setOutlineEnabled(false)" it will be re-shown when the alternate
action is deactivated (basicaly every time shift is released).
So I introduced a new "setOutlineVisible" that can be used to temporarilly hide the
outline. "setOutlineEnabled" is still present and permanently hides the outline,
regardless of what is set via "setOutlineVisible"