Rework text-rendering.
This commit implements the 'text-rendering' property to select different
kinds of considerations when drawing the text. It can get a DPI appropriate
value through use of hacks. Furthermore, it makes sure that gradients use
the full textShape bounding box for drawing.

It also allows using bitmaps for rendering if there's no outline.

Finally, it fixes KoClipMaskPainter to work when both shape and mask alpha
values involved are semi-transparent.