Automatically set KisColorInput height using layout
Previously, the height was determined using the textbox's sizeHint().
As discovered by deif_lou in 453146, the font in the docker may be
adjusted later, this renders the size hint inaccurate if the component
was created in a constructor before the font size change was applied.

Set the height using layout instead of manually setting a size hint.