Make tree indentation of Layers docker configurable
100% of the thumbnail size just takes too much space for many people,
leaving little space (or even none at all) for a layer name.
It now has a range of 20% to 100%, adjustable with a slider below
the thumbnail size slider.
While at it, remove a hack to update the view layout and slightly
optimize the delegate.

Also a few small quirks were fixed:
- the settings button did now update on theme changes
- the slider styling did not update on theme changes due to style sheets
- the thumbnail slider did not update the setting unless dragged with
  the mouse; mouse wheel and keyboard shortcuts failed to do so.

BUG:438457