Implement font shorthand, font-size-adjust and extra variation css.
Font-size-adjust was calculated wrong, so we need to avoid loading it when
it comes from an old kra file.

Font-variation-settings and font-optical-sizing are CSS-font-4 settings that
allow some extra control and are pretty widely implemented already.

And the font shorthand is my best attempt at parsing it. There's also
some initial work for overflow-wrap, which still needs implementing.