Fix text-decoration to use the visual order so it works with bidi text.
The original code broke when a piece of bidi text would go over a line-
break, which meant the text-decoration code didn't know where to start
drawing. By using the visual orde this is resolved.

This also greatly simplifies the text-decoration code, as well
a small fix to ensure that hanging punctuation before a collapsable
space is hung if the space is collapsed.