Cache, part 1: enable caching of FcPatterns and FcFontSets
Pango's documentation says that FcFontSets are cachable on their
patterns. Since we already create a pattern for every query, we can just
as well try caching both and use the former as a key (via FcPatternHash)
for keeping a tally of the font sets.

The need for using QHash and automated deletion makes Dmitry's resource
pointers unworkable; this commit replaces them with customized
std::shared_ptrs.