Set maximum width to the blending mode combobox on the toolbar
Before this commit, the blending mode combobox on the toolbar
was as wide as the widest item inside. It isn't needed, because
when unfolded, it shows all the items in a wider rectangle
which is wide enough for the widest item.

This commit sets the maximum width for the combobox, so it takes
less space on the toolbar than before.

If the user chooses a very wide item, the combobox
will just cut of the end of the item
(and if there are multiple blending modes that start with the same
string, and the ending is cut, the user can just click on the combobox
and will see the full text).