Set the correct colorspace on the internal color selector
This adds a separate instance of the internal color selector for
the background color selector, making sure that both fg and bg
selectors are handled the same way.

It also removes locking the colorspace on the internal color
selector: that was called only when setColorSpace was called
on the dual color button, which only happens if we switch
images in the view manager. Now it also sets the colorspace
of the current node, not the image.

BUG:452861