Make KisResourceLocator::saveTags use ResourcePaths
Before this commit, saveTags() used a custom function
that were reading the resource folder location
from configuration. This custom function wasn't used
anywhere else (it was used historically, but not
at the time this commit was made), and it didn't use the
`overrideAppDataLocation` functionality that ResourcePaths
have.

This commit removes the function altogether and changes
saveTags() to use ResourcePaths instead.