Create transform mask for clone and file layers
Automatically adds and selects transform mask for clone and file layer
when user attempts to use the tool on them.

The mask will be created only when user clicks the canvas, so no
accidental masks will be creating by just selecting the tool while
mentioned layers are active. Krita will also inform them with msg that
those layers can't be transformed directly and that new mask for them
was added.
User has to click again after first click, which is not ideal but not
sure how to go around that for now. Although, it doesn't create big
problem for the user it's just less smooth.

BUG:349152
Makes !510 obsolete.