Fix behaviour of toggle layer properties actions
Before this commit, toggle layer properties (visibility, alpha lock,
lock, and alpha channel disabledness) actions were only
setting all of the selected layers properties to the opposite
of the property at the current node.
This commit ensures that the actions actually toggle all layers
properties (so for example if you have two layers, A and B,
and A is visible, and the other is hidden, the action before
this commit would set both to the same visibility, and after this commit
will set A to hidden and B to visible).

This might need some discussion on whether this behaviour is
expected/desired.

BUG:459683