Fix calculating starting point for hyperbolas
The article was wrong, the hyperbola resulting axis are not always
horizontal, so it has to be checked and handled.
Moreover calculating one of the t0d/t0dd would be impossible,
so it has to be checked as well.

This commit also includes changing from qMax(x, y) to qMax(qAbs(x),
qAbs(y)) to normalize the original point correctly.

Changes were suggested by Mathias Wein and commited by Agata Cacko.