Implement snapWithLine
Some snap strategies only test x and y components, which leads to
issues when trying to snap a line extension, and said line is rotated.

This implements snap with line, which tries to find an intersection
or otherwise snap directly on the end of the line closest to the mousepos.