svgtexttool: Render preview when dragging to create text
This is achieved by creating a text shape for the preview, repositioning
and resizing (inline-size) as needed, and painting the text shape
directly as part of the tool decoration. This has some odd effects:

* The preview is rendered in the view resolution, making it a bit weird
  when the zoom level is high.
* The preview is overlaid on top of the image, i.e. does not respect the
  layer stack.
* The preview is rendered even outside the canvas region.

The preview shape is also directly reused to be added to the shape layer
once the dragging operation is finalized.