Change the layout of Reference Tool Options docker
Before this commit, the layout of the Reference Tool Options
docker was confusing, because it contained 7 icons (with one
actually invisible). This commit:
- splits the tool buttons into two groups, "Add" category
and "Everything else" category
- changes the "current layer" and "all layers" options
to the ones used in Fill Tool etc., the "copy" icon,
despite being used in "copy current layer", wasn't descriptive
enough
- changes the "plus" button into "folder" button, because all four
buttons in that category create/add new reference images, so the first
button isn't unique in that. The folder button shows that it allows
the user to add a reference image from a file.

Another possibility was to add plus to add icons used in the Add
category, but that would need creating new icons.