Change text on the Delete All button in Reference Tool
Before this commit, the button had a trashcan icon and text saying
"All". It was probably one of the only times where the meaning of
the button had to be constructed by both the icon and the text
(trashcan = remove, + text = all, => remove all).
This commit adds the word "Delete" which makes the button more clear.