Make Krita use .krf suffix by default for reference images
Before this commit, if you tried to export a set of reference images
and just pressed ok, or wrote the filename without a suffix
(which is not well known for users), krita will complain about
the empty suffix. This commit ensures that the default mimetype
is known so Krita will add the suffix if the user doesn't do it.