Implement CSS Palette support
Add the possibility to import CSS palette in Krita, in the same way
as the .gpl ones. Those palettes can be generated by GIMP or created
by the user, but the only property considered is "color".

The following color format is supported (based on their CSS syntax) :
    - rgb
    - rbga
    - hsl
    - hsla
    - hexadecimal

The file's suffix must be ".css".
If there is an error in the syntax or the color definition, errors
will be raised and the palette loading aborted.

FEATURE: [367694]