Make Freehand Path Tool remember options
Before this commit, Freehand Path Tool didn't remember any options
set by the user. This commit adds configuration entries remembering
those options (like what mode is used or the angle or whether to
optimize the stroke).

It also includes a fix for a default cofiguration, where single step and
default value were mixed up. It DOES change the default setting for the
tool but it shouldn't be too much of an issue for users, and the new
default values make more sense.